Allow the results of schedule queries to be saved as point in time data
This functionality can only be used on non-privileged data sources.
This also required a change to oid type mappings; previously, the
columns info recorded in the data field of the query_results table had
null type for varchar fields. The change also necessitated bumping
snowflake-connector-python as an issue in that library was causing
requests to AWS Firehose to fail
(https://github.com/snowflakedb/snowflake-connector-python/issues/2).

fixup! Allow the results of schedule queries to be saved as point in time data